requirements.txt
streamlit==1.38.0
ultralytics==8.2.77
opencv-python-headless==4.10.0.84
tensorflow-cpu==2.16.1
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
